--- src/main/java/br/com/unicuritiba/projetoathus/domain/dto/SetAtivoDTO.java
@@ -1,5 +1,0 @@
-package br.com.unicuritiba.projetoathus.domain.dto;
-
-public record SetAtivoDTO(boolean ativo) {
-    
-}
