--- src/main/java/br/com/unicuritiba/projetoathus/domain/dto/SetNivelDTO.java
@@ -1,4 +1,0 @@
-package br.com.unicuritiba.projetoathus.dto;
-
-public record SetNivelDTO (Short nivel) 
-{}
