+++ src/main/java/br/com/unicuritiba/projetoathus/domain/dto/usuario/SetAtivoDTO.java
@@ -1,0 +1,5 @@
+package br.com.unicuritiba.projetoathus.domain.dto.usuario;
+
+public record SetAtivoDTO(boolean ativo) {
+    
+}
