+++ src/main/java/br/com/unicuritiba/projetoathus/domain/dto/usuario/SetNivelDTO.java
@@ -1,0 +1,4 @@
+package br.com.unicuritiba.projetoathus.domain.dto.usuario;
+
+public record SetNivelDTO (Short nivel) 
+{}
